deep optimization (read desc)
overhead (number/s is byte/s).

class, struct = [before=9][now=1]
list, array = [before=9[]now=3] (support max. of 65535 elements),
BigInteger = [before=5][now=3]

+++ src/src/primitive/interfances/IPrimitive.cs
@@ -4,7 +4,7 @@ namespace Byter
 {
     internal interface IPrimitive
     {
-        int Position { get; }
+        long Position { get; }
         bool IsValid { get; }
         IPrimitiveAdd Add { get; }
         IPrimitiveGet Get { get; }
